# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Siuji TOEFL exam application.",
        "title": "Siuji API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/participant/periods": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get periods the logged-in participant is registered for.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ParticipantExam"
                ],
                "summary": "List my exam periods",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.ParticipantPeriodListResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/participant/periods/{period_public_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ParticipantExam"
                ],
                "summary": "Get my exam period detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Period public ID",
                        "name": "period_public_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ParticipantPeriodDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            }
        },
        "/participant/periods/{period_public_id}/answers": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Can be called repeatedly while the exam is in progress — upserts by question.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ParticipantExam"
                ],
                "summary": "Save/update an answer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Period public ID",
                        "name": "period_public_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Answer",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SaveAnswerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.SaveAnswerResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            }
        },
        "/participant/periods/{period_public_id}/result": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the final score, section breakdown, and certificate URL (generated on first request if passed).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ParticipantExam"
                ],
                "summary": "Get exam result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Period public ID",
                        "name": "period_public_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ExamResultResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            }
        },
        "/participant/periods/{period_public_id}/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Marks the exam as started and returns all sections/questions (without correct answers).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ParticipantExam"
                ],
                "summary": "Start exam",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Period public ID",
                        "name": "period_public_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ExamSessionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            }
        },
        "/participant/periods/{period_public_id}/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Finalizes the exam, computes section scores and the final TOEFL score.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ParticipantExam"
                ],
                "summary": "Submit exam",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Period public ID",
                        "name": "period_public_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.SubmitExamResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ExamOptionResponse": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                },
                "option_text": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "public_id": {
                    "type": "string"
                }
            }
        },
        "model.ExamQuestionResponse": {
            "type": "object",
            "properties": {
                "audio_url": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ExamOptionResponse"
                    }
                },
                "passage": {
                    "type": "string"
                },
                "public_id": {
                    "type": "string"
                },
                "question": {
                    "type": "string"
                }
            }
        },
        "model.ExamResultResponse": {
            "type": "object",
            "properties": {
                "certificate_url": {
                    "type": "string"
                },
                "final_score": {
                    "type": "integer"
                },
                "min_passing_grade": {
                    "type": "integer"
                },
                "period_title": {
                    "type": "string"
                },
                "section_breakdowns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SectionBreakdownResponse"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.ExamSectionResponse": {
            "type": "object",
            "properties": {
                "position": {
                    "type": "integer"
                },
                "public_id": {
                    "type": "string"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ExamQuestionResponse"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.ExamSessionResponse": {
            "type": "object",
            "properties": {
                "period_public_id": {
                    "type": "string"
                },
                "sections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ExamSectionResponse"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "model.ParticipantPeriodDetailResponse": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "month": {
                    "type": "string"
                },
                "participant_status": {
                    "type": "string"
                },
                "public_id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "model.ParticipantPeriodListResponse": {
            "type": "object",
            "properties": {
                "period": {
                    "$ref": "#/definitions/model.PeriodSimpleInfo"
                },
                "public_id": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "model.PeriodSimpleInfo": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "public_id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "model.SaveAnswerRequest": {
            "type": "object",
            "required": [
                "option_public_id",
                "question_public_id"
            ],
            "properties": {
                "option_public_id": {
                    "type": "string"
                },
                "question_public_id": {
                    "type": "string"
                }
            }
        },
        "model.SaveAnswerResponse": {
            "type": "object",
            "properties": {
                "option_public_id": {
                    "type": "string"
                },
                "question_public_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.SectionBreakdownResponse": {
            "type": "object",
            "properties": {
                "correct_count": {
                    "type": "integer"
                },
                "scaled_score": {
                    "type": "integer"
                },
                "section_title": {
                    "type": "string"
                }
            }
        },
...
        "model.SubmitExamResponse": {
            "type": "object",
            "properties": {
                "period_public_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "submitted_at": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.ResponseNoData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "TempToken": {
            "description": "Type \"Bearer\" followed by a space and temp JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}